+++ Assets/Code/Game Systems/Resources System/Indicators/Indicator.cs
@@ -3,14 +3,14 @@ using UnityEngine;
 
 public abstract class Indicator
 {
-    [SerializeField] private int current;
-    [SerializeField] private int max;
+    [SerializeField] private float current;
+    [SerializeField] private float max;
     [SerializeField] private IndicatorView view;
 
-    public int Current => current;
-    public int Max => max;
+    public float Current => current;
+    public float Max => max;
 
-    public Indicator(int current, int max, IndicatorView view)
+    public Indicator(float current, float max, IndicatorView view)
     {
         this.current = current;
         this.max = max;
@@ -19,17 +19,17 @@ public abstract class Indicator
         view.Init(current, max);
     }
 
-    public void Increase(int value) 
+    public void Increase(float value) 
     {
         current = Math.Min(max, current + value);
         view.ChangeValue(current, max);
     }
-    public void Decrease(int value) 
+    public void Decrease(float value) 
     {
         current = Math.Max(0, current - value);
         view.ChangeValue(current, max);
     } 
-    public void ChangeMax(int value) 
+    public void ChangeMax(float value) 
     { 
         max = Math.Max(0, value);
         view.ChangeValue(current, max);
@@ -39,17 +39,17 @@ public abstract class Indicator
 [Serializable]
 public class Health : Indicator
 {
-    public Health(int current, int max, IndicatorView view) : base(current, max, view) {}
+    public Health(float current, float max, IndicatorView view) : base(current, max, view) {}
 }
 
 [Serializable]
 public class Mana : Indicator 
 {
-    public Mana(int current, int max, IndicatorView view) : base(current, max, view) {}
+    public Mana(float current, float max, IndicatorView view) : base(current, max, view) {}
 }
 
 [Serializable]
 public class Stamina : Indicator 
 {
-    public Stamina(int current, int max, IndicatorView view) : base(current, max, view) {}
+    public Stamina(float current, float max, IndicatorView view) : base(current, max, view) {}
 }
